requirements.txt
requests==2.32.3
python-dotenv==1.0.1
openai==1.37.1
fastapi==0.68.0
uvicorn==0.15.0
